consolidate duplicated time functions
